+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-- `awless whoami` now returns your identity, attached (i.e. managed) and inlined policies
+- `awless whoami` now returns your identity, your attached (i.e. managed), inlined and group policies
 - Rudimentary security groups port scanner inspector via `awless inspect -i port_scanner`
 - Template: compile time check of undefined or unused references
 - [#78](https://github.com/wallix/awless/issues/78): Show progress when uploadgin object to storage